Fix Room Class
