+++ Enums.java
@@ -1,0 +1,21 @@
+public class Enums{
+    public enum Equipment{
+        NONE,
+        SMART_BOARD,
+        WHITEBOARD,
+        COMPUTERS,
+        PROJECTOR
+    };
+
+    public enum RoomType{
+        NONE,
+        OFFICE,
+        MAJLIS,
+        G_STUDY,
+        S_CASES,
+        M_CASES,
+        L_CASES,
+        AUDITORIUM,
+        LAB
+    };
+}
