--- enums/Equipment.java
@@ -1,8 +1,0 @@
-package enums;
-public enum Equipment {
-    NONE,
-    SMART_BOARD,
-    WHITEBOARD,
-    COMPUTERS,
-    PROJECTOR
-}
