--- enums/RoomType.java
@@ -1,12 +1,0 @@
-package enums;
-public enum RoomType {
-    NONE,
-    OFFICE,
-    MAJLIS,
-    G_STUDY,
-    S_CASES,
-    M_CASES,
-    L_CASES,
-    AUDITORIUM,
-    LAB
-}
